[split] make base Dtab used in BindingFactory a Stack.Param
Problem:

The base Dtab used in BindingFactory is hard-coded to the global
Dtab.base; in rare cases it needs to be changed per client.

Solution:

Make the Dtab a Stack.Param and add a configuration for it in
ClientBuilder.

Result:

The base Dtab may be set per client.

RB_ID=446719