[split] Finagle: clear locals before entering server dispatch loop
Problem

Locals from the caller of Server.serve* (alternatively, ServerBuilder.build)
can set locals that are propagated to the server's dispatch loop. This
is not usually what we want. The server should start with a clean slate.

Solution

Introduce the ability to clear locals. Do so when entering dispatch loops.

RB_ID=450639