[split] Transport.copyToWriter
Problem

Creating a Reader from a Transport is complicated and easy to get
wrong.

Solution

Transport.copyToWriter takes a Transport[_, A] and a function to
serialize `A` to `Buf`, then continuously reads from the Transport
writing the output of the serializing function to the Writer.

Result

A simple and reusable method to get a Reader from Transport provided
there is a mapping between the output type of Transport and `Buf`.

RB_ID=396003