[split] finagle-core: include name resolution in tracing request span
Problem:

Currently name resolution happens outside the tracing span of a client
request, so there is no way to see if a request is slow or failed
because of slow resolution. (Verily it happens entirely outside of stack
construction, so it's not even subject to the service acquisition
timeout.)

Solution:

1. Make FactoryToService a Stackable module, and put it into the default
StackClient stack above TraceInitializerFilter so service acquisition is
included in the request span. (This is necessary because name resolution
happens in service acquisition.)

2. Make BindingFactory a Stackable module, and move it into the default
StackClient stack above FactoryToService so name resolution is included
in the request span.

Result:

Tracing from name resolution is included in the request span.

RB_ID=471921