finagle-core: Enable SNI by default for finagle clients over ssl
Problem

finagle's implementation of SSL does not use Server Name Indication
(SNI).

Solution

finagle must use SSLEngine, not SSLSockets, because SSLSockets have
mostly blocking operations.  In order to use SNI with SSLEngine,
we must supply a peerHost and peerPort to SSLContext.  This means
threading a host and port through to SSL connection establishment
time.  The only reference that we have to the downstream is the
SocketAddress, so we use that by default--however, if the client
specifies a hostname to verify against, we override the host with
that one.

Result

SNI is automatically turned on for all clients that enable TLS.
When host verification is also turned on, we use the same host for
host verification and SNI.

RB_ID=481449