[split] finagle-core: Introduce a 'parameter' field in Stack.Head
Problem

Stacks are inherently stateless and thus shared between
finagle clients and servers. The way we inferred the
parameters that a module used to configure itself
attached state to a Stack and violated this guarantee.

Solution

Introduce a `parameters` field in Stack.Head which servers
as documentation and can be dynamically introspected to
determine the fields and values that a module is interested
in. To make it easy to do the right thing, abstract definitions
for `Stackables` now take type parameters that map to the
parameters a module is interested in.

Result

We can now collect interests per-module:

val stk = ...
val params = ...

val parameters = for {
 n <- stk.tails
 interest <- n.head.parameters
} yield params(interest)

RB_ID=484747