finagle-http: remove content length headers when chunking
Problem

Content-Length headers are not removed by Netty's compressors. This
causes content which is streamed, chunked, compressed, and also has a
Content-Length header to respond with the wrong value.

Solution

Always remove content length headers when streaming replies.

Discussion

This is a bad violation of modularity. Arguably the issue lies in
Netty here, and we should follow up.

RB_ID=496263