Add documentation about tries stats scope
Problem

The fact that the /success and /failure stats don't take retries into account
can lead to confusion, and the /tries scope isn't documented very well.

Solution

Add a paragraph about /tries and measuring success rate to the metrics section
of the users' guide.

Result

Less confusion about how to measure success accurately for clients that are
configured to retry.