Fix TraceId equality
Problem

TraceId cannot be used as a hash key as the hashCode is the default
hashCode of a case class which is based on all fields. However the
equals method only takes into account traceId, parentId, and spanId.

Solution

Override the hashCode method in TraceId to only take into account
traceId, parentId, and spanId.

Result

TraceId can be used as a key for the hash map of buffered spans.

RB_ID=504661