finagle-core: Made ServerChannelConfigurationTest more patient
Problem

"org.scalatest.exceptions.TestFailedException: Expected exception
com.twitter.finagle.ChannelClosedException to be thrown, but
com.twitter.finagle.ChannelWriteException was thrown."  This seems
to indicate that we expect to get a ChannelClosedException but we
actually weren't able to start writing bytes over the wire because
we were still establishing a connection.

Solution

Made it more patient.

RB_ID=504977