Finagle-memcached cas() to match on NotFound()
Problem:
TimelineService has discovered that memcached's cas() operation can return a NotFound() state when the key being set no longer exists. This is not checked for in the finagle client and will instead throw an IllegalStateException.

Solution:
Add the case matching NotFound() which will return false for the operation.

Results:
For cas operations in which the key has been deleted between operations, it will return false. This matches current understanding that the existing key attempting to be written has been changed since last check.

RB_ID=503161