finagle-core: distribute traffic according to NameTree.Union weights in BindingFactory
Problem:

We would like to distribute traffic across several service
clusters (each having an independent load balancer) according to the
weights in a NameTree.Union, but BindingFactory does not respect them;
rather it evaluates a tree containing Unions down to a single set of
endpoints and passes the whole set to a single load balancer.

Solution:

Evaluate a NameTree into a tree of ServiceFactories, where a Union
becomes a ServiceFactory which distributes traffic according to the
weights in the Union, and a Leaf become a ServiceFactory which balances
over the endpoints contained in the Leaf's Name.Bound.

RB_ID=498167