util, finagle: introduce c.t.util.events.Sink
Problem

There is no centralized place where runtime events are sent
making it difficult to diagnose issues.

Solution

Introduce the `com.twitter.util.events.Sink` trait so that
arbitrary events can be sent there enabling further analysis.

An emphasis is placed on runtime performance and avoiding
object allocations in order to allow this to be used in
high throughput environments.

The first provided implementation of Sink is `c.t.u.events.SizedSink`
and it implements a fixed size, in-memory, ring buffer of events.

To test, this can be enabled is via the
`com.twitter.util.events.sinkEnabled` flag and sized via
`com.twitter.util.events.approxNumEvents`. It is then exposed to your
application at `Sink.default` and is as of now wired up into
`MetricStatsReceiver` in finagle-stats and `SamplingTracer` in
finagle-zipkin.

Result

Benchmarks of `SizedSink.event` show 0 bytes allocated and
~500 nanoseconds per call with 8 threads and a capacity of 2^20
(on my 2.8 GHz Intel Core i7 laptop).

This will enable future work to expose information in a twitter-server
admin endpoint as well as wiring up logging to a `Sink`.

RB_ID=503483