[finagle-http] fix http server error handling
This commit moves the behavior of CheckRequestFilter to the
HttpServerDispatcher. The dispatcher now returns appropriate
HTTP response codes depending on the error condition and
forces closure of the unusable connection.

Previously any exception thrown by the decoder would be converted
into a 400 Bad Request response. With this commit, exceptions
caused by exceeding size limits are turned into 414 or 431 as
appropriate.

Also, the keep alive header is set to false on the response so
that the unusable connection is closed, preventing subsequent
requests from failing.

RB_ID=510449