finagle-mux: Introduce Mux request and response types
Problem:

It is not possible to write a generic, transparent mux proxy because destination information is not passed through with requests.

Solution:

Change the Mux api to be Service[mux.Request, mux.Response] instead of Service[ChannelBuffer, ChannelBuffer] so that destination Paths are included with Mux requests. Furthermore, these types expose Bufs instead of ChannelBuffers to support future netty changes.

RB_ID=502005