finagle-http: Keep clients and servers in sync with regard to closed connections
Problem

Client and Servers become out of sync about whether to close the
connection because a client may send a request that is HTTP/1.1 but the
safe codec may end up converting it into a HTTP/1.0 request. This
results in the next client request failing with due to the connection
being closed.

Solution

Set the Connection header for a response only when the server wants to
notify the client of connection closure. This implies that the default
"keepalive" behavior will be applied based on the protocol version.

Result

Bad HTTP requests due to long URIs will cause the client to close the
connection, just like the server.

RB_ID=518239