finagle-mux: Allows hosts with static ips to be perceived as alive again
Problem

On machines with dynamic static IPs, a fast restart means that they
will never get kicked out of the LB, and be re-constructed.  This
means that it's not OK to set them to be unavailable permanently.

Solution

(Temporarily) Make LeasedFactory "available" again once the underlying
service is dead.

RB_ID=522401