[finagle-core] support bound residual paths
Problem:

When a Namer binds a Name, there is no mechanism for "residual" (unconsumed) path components to be used by the client.

Solution:

Add a 'path' field to Name.Bound. Change the BindingFactory module to support request filtering based on this residual path. This anticipates further changes to namers (to provide bound residuals) and to mux (to inject residual paths into the downstream request).

RB_ID=502809