[finagle-http] support HTTP/1.0 keep-alive
Problem

The connection management behavior in finagle-http servers
is divided between the finagle-based "close()" mechanism and
the http-based "Connection" header. As a result, the keep-
alive behavior of the connection is dependent on the HTTP
protocol version of the messages being exchanged.

Solution

This commit makes the finagle-based "close()" mechanism
authoritative. Connections are kept alive for as long as the
client is capable and willing, unless close is called on the
server.

Result

As a result, we now support HTTP/1.0 keep-alive connections.

RB_ID=524423