finagle-stats: Introduce ImmediateMetricsStatsReceiver
Problem

By default, MetricsStatsReceiver uses WindowedApproxHistogram to report stats on a sliding
window of 60 seconds and with a delay of at most 20 seconds.
This is useful in production but annoying when you want to write tests because you need to
wait before the data will be aggregated.

Solution

Refactor the MetricsStatsReceiver constructor, which now takes a histogram builder parameter.
Introduce a new class ImmediateMetricsStatsReceiver, which uses ApproximateHistogram under
the hood (without any window).

RB_ID=530261