Update sbt projet definitions of (util, ostrich, scoorge, twitter-server)
Problem

In order to release our libraries for the open-source community, we need to update
the version of those.

Solution

Update the version in the sbt project definition.
While I was there, update also sbt to the latest version and increase the jvm
configuration for the code cache.

RB_ID=537603