[finagle-httpx] Remove deprecated code, shrink access
Problem

Deprecated code in http was copied over when we created httpx.
This is a good time to get rid of it (i.e. before people start
using httpx directly), and also shrink the access scope of
things that should be private to httpx/finagle.

Solution

Remove deprecated code and shrink access to internals.

Result

Leaner meaner httpx.

RB_ID=539076