Complete migration of finagle-redis from specs to ScalaTest
Problem

Jeff Davis migrated the majority of finagle-redis's tests from specs to
ScalaTest, but there are a few remaining gaps and some formatting issues.

Solution

This commit completes the migration, changes multi-line asserts (with separate
definitions for actual and expected values) to inlined asserts, and removes
remaining specs tests and dependencies. It removes the outdated (and unused)
`BtreeClientSpec` without a ScalaTest replacement. It also adds ignore
annotations to all integration tests and updates the filter in the SBT config.

Result

Finagle does not depend on specs.