[util-core]: Safeguard TimeFormat from misuse of "week year" ("YYYY")
Problem

It is easy to mistakenly use "YYYY" instead of "yyyy" in a date pattern string.
The only proper use of week year is in concert with the week number, e.g.
"YYYY-ww".

Solution

Implement a java.text.DateFormat that checks the pattern on initialization,
and use it from TimeFormat.

Result

Any format attempting to format non-week dates with the week year will
fail on construction.

RB_ID=552028
TBR=true