[finagle,t-s-i] support bound residuals on serverset names
Problem:

The serverset namer treats the whole name as a zk path, and so it cannot handle residual paths.

Solution:

Introduce a new twitter-internal serverset namer, com.twitter.fixedserverset,
that expects serversets to be named in the form:
  /HOST/twitter/service/ROLE/ENV/JOB[:PORT]

Names that do not meet this expectation are resolved negatively (and will fall
back to the old com.twitter.serverset namer that uses the full path and does
not support residuals).

RB_ID=535469