[trivial]: Use bufSize in streamChunks recursive case
Problem

streamChunks takes an optional bufSize argument, but it discards it in
the recursive case. This means bufSize is respected in only the initial
call to read(), subsequent calls default bufSize to Int.MaxValue.

Note: This bug is impossible for the user to encounter because of the
access of Reader.streamChunks (i.e. only http/x packages can use it).
All call sites omit the bufSize argument, and so the recursive case has
the correct bufSize. However, it's possible to run into this issue by
packaging code under the http/x namespace, thereby subverting the access
barrier to call Reader.streamChunks directly.

Solution

Use bufSize in the recursive case.

RB_ID=574416