Target move database should be an integer string
Problem

Redis integration tests are failing with an exception. Usage of the "move"
command expects an integer string as the second parameter, but we give it
alphabet characters. Redis' response to this is undefined.

Solution

Changed the second parameter to the `toDb` index, which was probably the
original intention.

Result

Integration tests now pass.