finagle-core: improve name resolution tracing
Problem:

In some cases tracing on name resolution doesn't work as desired: it
correctly traces successful lookups and lookups which throw exceptions
during naming, but does not correctly trace Neg/Fail lookups (which
produce no Name.Bound and never reach the NamerTracingFilter in the
stack, but do not throw exceptions) or exceptions thrown in service
acquisition such as connect timeouts (because the NamerTracingFilter is
effectively on the Service not the ServiceFactory).

Solution:

Trace the NameTree resulting from name resolution (so Neg/Fail lookups
are traced), and use ServiceFactoryProxy's to record tracing in the
service acquisition phase (so tracing still happens when there is a
service acquisition exception); consolidate tracing in BindingFactory
and test all these cases.

RB_ID=575210