[util-collection] change semantics of RecordSchema: all fields are mutable until locked Problem -------
The semantics of immutable fields with default values are surprising to users:

  // declare a field with a default value
  val schema = new RecordSchema
  val field = schema.newField(42)

  // create a new record, and access the field (not yet present in the record)
  val record = schema.newRecord()
  val value = record(field)

  // try to associate a value with the field in the record
  record.put(field, 31337) // throws IllegalStateException, because `record.apply` and `.get` memoize the default in the record

That is, both reads and writes are treated as initialization.

Solution
--------

Remove the notion of an immutable field, and replace with a new concept of locking. Fields are mutable until they have been explicitly locked.

Additionally, remove .get(): accessing an uninitialized field should have the same dire consequences it would in C. If the user needs an optional field, they must specify Field[Option[...]] and initialize it to None.

Result
------

We can stop arguing about what the semantics of defaults should be.

RB_ID=576168