finagle-memcache: KetamaFailureAccrual checks failureCount before failing immediately
Problem

There is a bug in KetamaFailureAccrualFactory, when the underlying
service is in a busy status, it fails immediately without checking
accumulated failure counts.

Solution

Check the state of KetamaFailureAccrualFactory instead of status,
since busy value of the latter can be inherited from the underlying
serviceFactory, while the former is set purely based on failureCount.

RB_ID=576904