Problem
ThriftClientBufferedCodec class extends ThriftClientBufferedCodec,
so when we call ThriftClientBufferedCodec.get() it actually
returns ThriftClientFramedCodec

Solution

Implement get function in the ThriftClientBufferedCodec
object so that it returns ThriftClientBufferedCodec.

Result

You should be able to call ThriftClientBufferedCodec.get() and
return a correct value, which is very useful for Java development.

RB_ID=583920