util-app: Adds Flags to registry
Problem

Flags with interesting values should be added to the registry.

Solution

There is a solution in a few parts.

1.  How to hook into flags?

The simplest solution is to do it at read time.  Reading is done
in a few different places, but the simplest way is to just put it
in the getValue method of Flag, which works for both GlobalFlags and
for regular Flags.

2.  Which registry to register in?

The registry is global, and it seems somewhat silly to program as
if it's not, so I've taken the same approach as we do with Time.

Result

Flags are added to the registry after read time, and are structured
like so:

/registry/flags/foo = true

RB_ID=587044