util-stats: Deprecates StatsReceiver.time{,Future}
Problem

StatsReceiver#time{,Future} encourage the bad habit of creating a
new Stat instance every time you time something.

Solution

Instead of letting a method hide the Stat creation, require that
the Stat be passed in explicitly.  This was done in an earlier
review, this review deprecates the StatsReceiver#time{,Future}
methods and replaces most of the uses in birdcage and science.

Result

StatsReceiver#time{,Future} are no longer widely used in the twitter
code base, and we have created a path to removing the
StatsReceiver#time{,Future} methods.

RB_ID=588854
TBR=true