finagle-core: move RequeueingFilter below BindingFactory
Problem:

In cases where Union nodes appear in a NameTree, we'd like requeues to
hit the same destination cluster rather than being distributed across
the Union. We are introducing Union nodes in place of multiple client
stacks (with external distribution across stacks), so making requeues
hit the same cluster retains the existing behavior, stats, etc.

Solution:

Move RequeueingFilter below BindingFactory. It's also necessary to move
FactoryToService (since requeues are mostly on service acquisition),
Role.prepFactory (currently used to delay closes for streaming HTTP
responses; it must go below FactoryToService), TimeoutFactory (must go
below FactoryToService to be effective), and Role.requestDraining (also
must go below FactoryToService to be effective).

We additionally add a FactoryToService and TimeoutFactory above
BindingFactory, because we want to rebind on each request (to account
for Dtab.local) and protect binding with a timeout; name binding happens
in service acquisition.

A higher-level point of view on all of this is that BindingFactory
should be at the top of the stack; the only thing outside it is tracing,
so that binding is captured in the request span.

RB_ID=600920