finagle-core: fix typing of StackClient.configured
Problem:

StackClient.configured is a StackTransformableClient, not a StackClient.

Solution:

Override the configured methods to the more specific types. Along the
way, rename StackTransformableClient to StackBasedClient, and clean up
the typing: separate the transform method into a Stack.Transformable
trait, and where we need more-specific types for Parameterized and
Transformable methods, mix in more-specific instantiations of the
traits.

RB_ID=603550