finagle-{stats,ostrich4}: Use httpx instead of http
Problem

We are migrating to finagle-httpx from finagle-http and
as part of that we need to fix our internal dependencies
to avoid using finagle-http.

Solution

Switch to using finagle-httpx and have twitter-server serve
the metrics endpoints properly as well.

twitter-server now depends on both finagle-http and -httpx
as a bridge until it is migrated to just using httpx.

RB_ID=604718