Added ability to add to the ignored packages sequence for LoadService
Problem

The problem being fixed here is something we ran into while working with finagle in our application similar to the issue posted in the following issue https://github.com/twitter/finagle/issues/309. In our case it was not the /sys directory however but an extremely huge FS mounted on the project root and LoadService was getting stuck descending into that directory structure everytime.

Solution

This pull request adds the ability to add additional packages/paths to the 'ignoredPackages' sequence in LoadService via a method called 'addIgnoredPaths' in LoadService. This method can take a single or multiple paths as String arguments and add them to the private sequence (which is now mutable).

Result

After this change, users would be able to customize the ignoredPackages for the recursive directory scan that LoadService performs (in several flows) and avoid getting stuck in directory structures that they do not want scanned.