[util] Use proxyTo instead of become
Problem

In the case where it is passed a promise, Promise.become merges the
listeners of two promises, an operation which requires modifying the
internal state of both Promises. The complication arises when the
promises each have an interrupt handler: one gets overwritten.

Consider the following, which demonstrates the behavior of `become` on
promises with interrupt handlers.

    val a, b = new Promise[Unit]
    a.setInterruptHandler { case _ => println("A") }
    b.become(a)
    b.setInterruptHandler { case _ => println("B") }
    a.raise(new Exception)

This prints "B", the action in the interrupt handler for `b`, which is
unexpected because we raised on `a`.

Solution

A cleaner solution is to use proxyTo, which does not merge listeners,
but relies only on the standard Future.respond. We can set an interrupt
handler on proxy, which will raise on the proxied future.

RB_ID=610718