Changed localhost to 127.0.0.1 to prevent ipv6 ip
Problem

When running the tests on my machine, which has both an ipv4 and 6 interface,
the tests fail because the endpoints are added to the load-balancer twice and
therefore the tests fail with the message that 2 != 1.

Solution

I've changed the hostname from localhost to 127.0.0.1 in the afflicted test. If
this is really the solution than there are probably more tests where tests where
this should be fixed. Additionally I added a better error message to the
assertion.

Result

Tests will run on my machine.