Bugfix memcached Zipkin integration.
Problem

The memcached Zipkin integration was broken. When getting a value from memcached
the cache hit/miss annotation was submitted after the cr annotation was submitted.
This had as a result that a new span was created with same id which was eventually
submitted by the DeadlineSpanMap once the deadline was exceeded. The span submitted
by the DeadlineSpanMap overrides the duration of the earlier submitted span and messes
it up. It also makes the zipkin-web trace overview almost unusable.

Solution

Instead of submitting the cache hit/miss annotations using the TraceInitializerFilter.role,
submit them using ClientTracingFilter.role before the cr annotation is submitted.
This means the memcached specific tracing filter submits following annotations:
cs -> memcache specific annotations -> cr. To be able to do this I updated the Codec trait
to be able to override filters which have a custom role and not the TraceInitializerFilter.role.

Result

The memcached span does not end up in the DeadlineSpanMap and the spans are correctly
submitted now including cache hit/miss annotations.  The duration is correct and the
trace overview in zipkin-web looks good again.