finagle-mux: FailureDetector closes connections on long inactivity
Problem

When a network partition occurs, the connection used by Mux may never be closed.
This is usually not a problem because Timeout/FailFast/FailureAccrual and others
will mark the status of the ServiceFactory as Busy. But for clients configured
with only one server, this never closes the connection.

Solution

Modify the FailureDetector, to close the underlying connection after a
(configurable) period of inactivity (no response to ping).
This period is configured as a multiple of the moving average of the ping latency.
By default, this is disabled but I recommend using a pretty large value (in the
range of 1000).

Modifications

I also introduced a new stats "failuredetector/ping_latency_us" which expose an
histogram of the ping latency in microseconds.

RB_ID=613666