Made RetryingFilter extendable.
Problem

The main logic of the RetryingFilter is good enough for me, however it does not
pass the request to the policy. In my case I always want to retry GET requests.
Therefore, I want to extend the RetryingFilter such that the new apply() method
can use the request to create a new policy. In the current situation that is not
possible because the dispatch method is private. Consequently, the
RetryingFilter is not extendable.

Solution

Changed the access modifier from private to protected.

Result

Classes that extend the RetryingFilter can now call dispatch.

Signed-off-by: Steve Gury <stevegury@twitter.com>

RB_ID=615154