Problem
Scala creates Functions instead of a PartialFunction when used
anonymously as in:

  Future.onFailure { case NonFatal(e) => ... }

This meant that the special handling introduced in RB_ID=588992 failed
to actually help. The unit test "lied" because it passed explicit
PartialFunctions instead of the more idiomatic anonymous ones.

Solution

Revert the changes introduced in RB_ID=588992 and fix the call sites
that were using it incorrectly.

RB_ID=620938