[finagle-core] Fail readq before close
Problem

Failures sent upstream from netty (via handleUpstream), don't fail the
readq quickly enough. The fail calls close(), which hands execution back
to netty (via Channels.close).

We observe in one case that after sending a failure, netty is still able
to send a subsequent message, which ends up on the readq, before we fail
it.

The related issue is that Netty sends upstream a partially deserialized
and corrupt HttpResponse after sending failure.  The ChannelTransport
should, upon receiving this failure, set the readq to Failing, which
causes rejection of subsequent offers. But because of the order in which
we call readq and close, the corrupt HttpResponse makes it into the
readq before it is set to Failing.

Solution

We fail the readq before close. This causes readq to reject any offers
after the receipt of the first failure.

Now, when Netty sends a failure and then the corrupted Response, the
client dispatcher fails instead of passing the bad Response back to the
user.

RB_ID=624140