finagle-mysql: transaction support
Problem

Currently it is not possible to run multiple statements using a
single transaction.

Solution

Add the transactional method that executes a function within a transaction.
The function's input is a client that has only one underlying database
connection with an open transaction. The provided client instance
must be used for all database interactions and never leak to outside
the transactional block scope.

The transaction is automatically committed in case of a success and
automatically rolled back in case of a failure.

Result

Users are able to run multiple statements using a single transaction.