finagle-core: Surface TLS peer cert info per request via Transport
Problem

While it's possible to enable 2-way TLS, there's no way of accessing any details of the SSL session on a per-request basis, which is useful client/server identification and auth.

Solution

SSL Session information is accessible via netty's SSLHandler, which is added to the ChannelPipeline for secure connections. This adds an optional SSLSession data member to the Transport trait. If present, this is propagated to each call to the service through Contexts.local via a dispatcher.

Result

SSL session info (if present) can be obtained by calling Transport.sslSession().

RB_ID=631870