finagle-mysql: Make PreparedStatment parameters type safe.
Problem

In the current Finagle MySQL, PreparedStatement#apply has a type
Seq[Any] => Future[Result]. This means that one might accidentally
end up passing things that cannot be written, and be in for surprises.
(Happened to us quite a few times now.) This might be revealed in
the warnings and the tests, but we can do much better: We can
eliminate such errors at compile time itself.

Solution

Change PreparedToStatement#apply to have type Seq[Parameter] =>
Future[Result]. Parameter is a class that makes use of existential
types and packs a value A and an implicit instance CanBeParameter[A]
together. CanBeParameter is a type-class. If you have an instance
of CanBeParameter[A] available, that serves as an evidence that,
well, a value of A can be a used as a parameter.

This PR ensures that all the types that were supported before are
still supported the same way.

null required special handling in a couple of places:

Parameter.wrap doesn't kick in for null because null is already a
member of Parameter type. So we need to sanitize the Seq[Parameter]
in ExecuteRequest#apply, and replace all nulls with NullParameter.
If there is a null with a static type A for which there is a
CanBeParameter[A] evidence available, then we might end up with,
for example, a stringCanBeParameter instance (if A is String) and
typeCode etc will yield wrong values. To avoid this, Parameter.wrap
has to guard for null and return a NullParameter in that case.  This
PR retains the support for null as is, for backwards compatibility.

Going forward, we could stop supporting null altogether and avoid
the complications I talked about before. We could instead support
Option which is a more principled way to deal with these sort of
things.

RB_ID=636592
TBR=true