twitter-zookeeper: remove eventually from Watcher tests.
Problem:
Watcher tests depend on eventually to assert changes in state, which can
make the test flaky if the system running the test is under heavy load.

Solution:
Set filters on conditions and use Await to allow the state to converge.

Result:
Watcher tests should be less flaky.

RB_ID=649858