[finagle-core] Introduce NameInterpreter
Problem
--

The Dtab and the name resolution mechanism are conflated: it is not
possible to resolve a name without first resorting to the Dtab, and the
Namer interface does not take the Dtab into consideration.  Resolution
must happen partly in the client.  This prevents outsourcing name
resolution and centralizing the Dtab into a separate service in a way
that retains existing use-cases.

Solution
--

Add an optional indirection, NameInterpreter, between (dtab, path) and
bound names, used by BindingFactory.

Result
--

We can override how dtab and path interact.

RB_ID=689726