Fix Travis CI and add code coverage for Finagle and related projects
Problem

This change addresses three closely related problems. The first is that our
Travis CI builds have been broken for a long time, in part because of the fact
that since we don't publish snapshots, the current development versions aren't
able to find their dependencies unless these are published locally. (Travis CI
also recently changed containerized builds to be the default, and we're still
unable to use the containerized build system with Finagle because of travis-ci
issue #3120, so I've turned this off for the relevant projects.) The second
issue is that we're not currently clearly indicating project status (whether a
project is retired, deprecated, or actively maintained). Finally, we want to
publish code coverage information for these projects.

Solution

This commit extends the work in Scrooge to provide a Travis CI script that
downloads and publishes dependencies locally. It also adds retries for the
dependency resolution and downloading step for each Travis CI config, as well as
caching for dependencies. It also adds a shields.io badge to the top of each
project README indicating the status of the project, and turns on Scoverage code
coverage reporting (and publication to Coveralls) for each project.

Result

If our Travis builds are failing, it's much more likely to be for legitimate
reasons. We also now clearly indicate code coverage stats and project status.

RB_ID=704444