finagle-core: Label FailFastExceptions for more useful log messages
Problem

It's difficult to know which endpoint is marked dead from the log
message of a FailFastException. While this can be inferred by observing
"marked_dead" counter increments in the same time range, adding the name
of the endpoint to the exception message makes it much easier.

Solution

Customize the exception message with a Label in the Stack scope.

RB_ID=716021