Introduce ServiceFactorySocketAddress so that Names may bind local services.
Problem:

When building a client using a Name--i.e. via Dtabs--there is no mechanism to
cause a client to bind to a local Service reference.

Solution:

Introduce a SocketAddress type that encapsulates a ServiceFactory.
StdStackClient's default endpointer bypasses the client's Transporter when such
an address is encountered.

Result:

Local services may be named.