Twitter-zookeeper: Make integration tests ready for ZK 3.5+
Problem:
Existing Apache ZK tests are locked to the API as it existed with ZK 3.4.

Solution:
Update test usage to use APIs available in 3.4 and 3.5+.

Result:
Tests works with 3.4 and 3.5+.

Note: Preallocation calls in ZKServerWrapper were added by tnarg@, who
said he added them in an attempt to see if there was a breaking point
in the server code. Reverting that change to keep things clean.

RB_ID=716700