Introduce server-side wire tracing.
Problem:

Clients annotate WireSend and WireRecv at the bottom of the finagle
stack.  There is no analog on the server-side (i.e. a top-of-the-stack
indicator).

Furthermore, it looks like WireTracingFilter.module was using the
ClientTracingFilter.role instead of WireTracingFilter.role. This may
cause client tracing issues?

Solution:

Introduce WireTracingFilter.serverModule at the top of the default
server stack.  This module installs a filter that annotates WireRecv
on receipt and WireSend on success.

Result:

Tracers have a definitive 'end of handling' marker.