finagle-httpx: reason phrase for status
Problem

The common HTTP statuses have an associated reason phrase.
It's convenient for applications to use this instead of the
toString of the case class.

Solution

Map the common statuses to reason phrases, exposed using
the method `reason` on the status.

Result

`println(Status.HttpVersionNotSupported.reason)` prints
"HTTP Version Not Supported" instead of the previous
"HttpVersionNotSupported".

For statuses without a common reason phrase, we fall back
to a generic phrase for the class of status (Informational,
Successful, Redirection, etc...).

RB_ID=722261