Problem
The problem being fixed here is something we ran into while working
with finagle in our application similar to the issue posted in the
following issue #309. In our case it was not the /sys directory
however but an extremely huge FS mounted on the project root and
LoadService was getting stuck descending into that directory structure
everytime.

Solution

This pull request adds the ability to add additional packages/paths to
the 'ignoredPackages' sequence in LoadService via a GlobalFlag
`com.twitter.finagle.util.loadServiceIgnoredPaths`. This flag can one
or more paths as String arguments and add them to the private sequence
(which is now mutable).

Result

After this change, users would be able to customize the
ignoredPackages for the recursive directory scan that LoadService
performs (in several flows) and avoid getting stuck in directory
structures that they do not want scanned.

RB_ID=724361