finagle-core: better stats and tracing of naming
Problem:

Latency stats and tracing around naming include service acquisition from
the underlying factory, so they are confusing / wrong.

This is in part because BindingFactory is structured as nested
ServiceFactoryCache's, with service acquisition at the bottom, but we
really only care about latency stats and tracing in DynNameFactory.

Solution:

Put latency stats and tracing in DynNameFactory instead of at the top
level of BindingFactory, and be careful that we are tracking only
naming, not service acquisition. Also, remove some tracing and stats in
ServiceFactoryCache which is more misleading than helpful.

RB_ID=730309