finagle-{core,serversets}: remove networkaddress.cache.ttl check for DNS poll / cache
Problem:

The networkaddress.cache.ttl property isn't a reliable way to check
effective DNS cache lifetime, but we use it in InetResolver to determine
the local cache lifetime and the polling interval; this can produce
stale DNS results.

Solution:

For serversets in ZooKeeper, it's OK to cache lookups permanently in
Finagle and never poll; we expect the membership to change by
re-registering a different host rather than by changing host -> IP
map. (Caching in Finagle is desirable because serversets are typically
large and change often; without a cache we'd lookup all hostnames in the
serverset on each change.)

For hostnames resolved directly we make no assumption that the host ->
IP map is fixed, we don't need to cache in Finagle (we can rely on
whatever caching goes on in the JVM or the OS resolver).

So we split these cases: for InetResolver we do no caching in Finagle
and poll on a 5-second interval; for Zk2Resolver we cache permanently
and do not poll. In neither case do we need to check
networkaddress.cache.ttl.

RB_ID=735006
TBR=true