util-stats: Improve StatsReceiver.toString
Motivation

The current implemenations of `StatsReceiver.toString` use the default
implemenation from `Object.toString`. As an example:

    "com.twitter.finagle.stats.StatsReceiver$$anon$1@1a6d1901"

This is not as helpful as it could be when looking at registries in
TwitterServer or in heap dumps.

Solution

Override `toString` for most of the standard `StatsReceivers`. This is
particularly useful when there is one or more `StatsReceiver.scope` in
use.

Result

More readable and useful output such as:

    "InMemoryStatsReceiver/srv/thrift"
    "MetricsStatsReceiver/clnt/auth"

RB_ID=740845