finagle/finagle-core: RetryingFilter can retry on successful requests
Problem
RetryingFilter allows retries on Throwables only. It would be useful to be able
to retry on successful responses that aren't sufficient.

Solution
RetryFilter now takes a RetryPolicy[(Req, Try[Rep])] to invoke on requests. The old
RetryingFilter, which took a RetryPolicy[Try[Nothing]], has been renamed to
RetryExceptionsFilter and is a subclass of RetryFilter.

RB_ID=732848
TBR=true