[Finagle-Memcached] Fix MockClient to also compute a casUnique value.
Without this, using the MockClient in consumers that rely on the `getResult` method directly
and the `casUnique` value in the returned `Value` instances will fail.

The `Interpreter` class was already using a `hash64` to compute a `casUnique` value. I extracted this to a companion object and
made it package-private, so it can be re-used for the MockClient. Then, I called the `hash64` method from `MockClient`
to compute the `casUnique` value.

RB_ID=741252