finagle-core: Export metrics from Netty's HashedWheelTimer
Motivation

There is currently no visibility into `DefaultTimer` which is a
service's primary `Timer`. This can help identify services that
may be blocking their timer or doing operations that may be better
offloaded to a separate thread pool.

Solution

  * Expose a stat that tracks the deviation from the expected time
    a task is expected to run at.

  * Expose a stat that tracks the number of pending tasks to be run.

Result

The fog is cleared. The metrics are exported at:

  "finagle/timer/deviation_ms"
  "finagle/timer/pending_tasks"

RB_ID=743233