finagle-core: Fix lifecycle of meanweight gauge
Problem

The meanweight gauge that is associated with a
`TrafficDistributor.Distributor` is not explicitly closed when weight
classes change. This can manifest as a form of a memory leak as the
gauges will not be garbage collected as fast as they should be.

Solution

Move the gauge to `TrafficDistributor` and explicitly manage
updates of the size on changes and remove it on close.

RB_ID=743315