Removes zipkin annotation.duration
The OpenZipkin project dropped support for Annotation.duration. This
field was rarely used, and when used often incorrectly. Moreover, the
zipkin query and web interfaces didn't support it. Dropping it
implicitly fixed bugs and confusion around the topic. This happened in
zipkin 1.9, and finagle-zipkin is the last trace instrumentation library
left to adjust.

See https://github.com/openzipkin/zipkin/pull/717