finagle.mux: make FailureDetector controllable via flags again
Problem:

When FailureDetector.DarkModeConfig was made the default, it was done in a way that broke the ability to overwrite the default behavior via the `com.twitter.finagle.mux.sessionFailureDetector` flag (the flag simply doesn't do anything now).

Solution:

Added "dark" as a potential flag value, set that as the default, and switched the default `Param` value back to `GlobalFlagConfig`.  This keeps `DarkModeConfig` as the default, but allows the default to be changed.

RB_ID=751701