finagle-core: Remove weights from c.t.f.util.Ring
Problem

We removed weights from the load balancer in a previous patch,
but there are some residual references to weights that needed
to be cleaned up.

Solution

Remove weights from c.t.f.util.Ring which is used by aperture.
This simplifies the construction of the aperture ring.

RB_ID=752149