[finagle-core, finagle-memcached] Add FixedInetResolver for clients or other resolvers that do not want to poll for DNS updates
Problem:

The InetResolver does not cache DNS lookups and polls every 5 seconds for updates per recent RB #735006. For certain scenarios this is not desireable as it can add hundreds of spurious lookups.

-- The zk2 resolver deals with large serversets and does not want to relook up all entries on every membership change. In this case, we do not expect changes to be communicated via changing ip addresses but instead host name changes.
-- The twemcached client uses a large fixed set of hosts serving different sections of the cache. These ip addresses should not be changing to indicate new servers in the set. Instead, it should change the hostnames.

Solution:

Allow clients to opt-into the "old" behavior of caching DNS lookups indefinitely. The zk2 resolver was already doing this. Promote the zk2 resolvers code to a new FixedInetResolver and have the twemclient builder specify its schema in its names.

RB_ID=753712