FIX: Race condition in IdleConnectionFilter
Problem:
If a connection is closed by the client during a request, `onClose`
could be filled (and the connection "removed" from the queue), before
the connection is added back to queue. This would make the
`IdleConnectionFilter` ineffective (It will always try to close a
connection that has already been closed).

Solution:
Verify if a connection have been closed before adding it back to the
pool of open connections.

Result:
`IdleConnectionFilter` still work even when a connection is closed
during a request.