finagle-core: DefaultMonitor respects Failure.logLevel
Motivation

Exceptions that propagate to the default `Monitor` became logged at
Warning level. This lead to excessive logging for `BackupRequestLost`.

Solution

`DefaultMonitor` is now the default `Monitor` and it respects
`Failure.logLevel`. This also necessitated a change so that
`BackupRequestLost` is no longer a singleton Exception. Instead, use
`BackupRequestLost.Exception`.

RB_ID=758056