[finagle-serversets] Zk2Resolver reports Addr.Pending when its ZkClient is unhealthy
Problem:

The Zk2Resolver relies on a healthy zookeeper client to accurately report updates to its clients, however, this resolver doesn't report any changes when its client dies or cannot connect.

Solution:

When the underlying zk client is unhealthy (not connected) for > 1 removal window as defined in the existing stabilization config, report that the Var[Addr] is no longer definitively Neg or Bound but rather Addr.Pending.

The finagle loadbalancer and traffic distributor already use cached values when receiving an Addr.Pending update from the resolver (and have a test in place to verify such). Thus this is a safe change to make for existing distribution scenarios.

Implementation:

 - ServiceDiscoverer exposes a health var.
 - HealthStabilizer buffers up unhealthy notices until a probation interval has passed
 - Zk2Resolver surfaces Addr.Pending if it receives unhealthy from its stabilized health var.

RB_ID=760771