finagle-core: Update StatsReceiver Link on Clients Page of User Guide
Problem:

The StatsReceiver link on the Clients page of the Finagle User Guide
points to StatsReceiver.scala within finagle-core. This file is now
located in util/util-stats.

Solution:

Change the StatsReceiver link in the documentation to point to the
util/util-stats version of the file on GitHub.

Result:

A working StatsReceiver link on the Clients page of the Finagle User
Guide.

RB_ID=764458