finagle: Introduce a budget for retrying failed requests
Motivation

Currently, clients can specify a `RetryPolicy` which determines which
types of failed requests are retried and how many retries to attempt.
This is flexible and gives users a good amount of control. However,
when downstream services are under duress, so called "retry storms"
can amplify the effects.

As an example, given a common setup of a front end talking to a middle
tier which talks to a backend. If both the front end and the middle
tier allow for 4 retries, and things go south on the backend, the
resulting amplification is 25x onto the backend service.

Solution

Introduce a dynamic budget that controls when it is okay to retry a
failed request such that whether or not a request can be retried is
now checked against both the policy as well as the budget.

The default budget `RetryBudget.apply()` allows for 20% of the total
requests to be retried on top of a minimum of 10 retries per second
to accomodate clients that have just started issueing requests or
clients that issue a low rate of requests per second.

Clients configured via `ClientBuilder` can customize the budget via
`ClientBuilder.retryBudget(Retries.Budget)` and clients configured via
`Stack` can use `configured(Retries.Budget)`.

RB_ID=760213