finagle-core: Fix race condition in FailureAccrual' revive timer task Problem
FailureAccrual's `reviveTimerTask` (the one we run periodically to check if
the dead host is back alive) has a race condition over its reference.

  1. [thread 1] Host goes down and we mark it dead for `x` seconds
  2. [thread 1] We run the timer task (`revivalTask`) to check if it's back alive in `x` seconds
  3. [thread 2] A `ServiceFactory` that wraps Failure Accrual calls `close` on it so we cancel the
     timer task (which is `None`) via `cancelReviveTimerTasks`
  4. [thread 1] Stores the reference to a timer task into `reviveTimerTask`

Since `cancelReviveTimerTasks` is not syncrhonized there is a race condition between 2 and 3 so there
is quite a chance we won't cancel the task seeing a `None` instead of `Some(task)` (memory leak).

Solution

Mark `cancelReviveTimerTasks` syncrhronized to make sure it will see the most recent value for
`reviveTimerTask`.

RB_ID=766119