finagle: Documentation Generation Cleanup
Problem

The CHANGES file has an incorrect layout for displaying release 6.29.0
and 6.30.0 within the documentation. Also, the pushsite script is
still looking for documentation generated based on Scala 2.10.

Solution

Modify the CHANGES file to use the correct format for release 6.29.0
and 6.30.0. Also modify the pushsite script to look for documentation
in the 2.11 folder, and add a note to the README regarding the
version used within the generated documentation.

RB_ID=771755