Add homepage and source reference to gemspec
Problem

There is no easy way to find the source code for the project from the
rubygems or the installed gem source.

Solution

Add the homepage and source_code meta data references to the gemspec.

Result

Both the homepage and the source_code urls are visible on rubygems.org
when looking at the finagle-thrift gem and accessible from the installed
gem source.