Add thrift dependency
Problem

This gem doesn't specify that it depends on thrift in it's gemspec. This
causes problems when then gem is installed and used because its
necessary dependencies are not installed and therefore when it is used
it throws an exception trying to load the thrift dependency.

Solution

Add the thrift dependency to the gemspec.

Result

When this gem is installed rubygems will make sure that the thrift gem
is installed as a dependency. This in turn will make sure that when
people use this gem that it doesn't throw an exception trying to load
the thrift dependency.