finagle-memcache Client.cas should differentiate between NotFound and Exists
The ConnectedClient does not differentiate between [Exists and NotFound responses](https://github.com/twitter/finagle/blob/develop/finagle-memcached/src/main/scala/com/twitter/finagle/memcached/Client.scala#L453-L455).
One could argue that these are both failures, however this presents a
problem when trying to expose the client as a service, as it's a lossy
representation to the caller.

I've added a cas2 method that returns a CASResult for each of the three
states: Stored, NotFound, and Exists. I've re-written the cas method in
terms of cas2, and fixed all of the subclasses of BaseClient where
people were redefiniting cas.

There will be a cas2 method that will expose the difference between a
NotFound and Exists response. All existing API consumers will be
unaffected.

RB_ID=771779