finagle-core: Add ServerDispatcher and AnnotatedServerDispatcher wrapper types.
Problem:

Clients annotate WireSend and WireRecv at the bottom of the finagle
stack. There is no parallel for a Server. In addition, we would like
to have a simple type for transforming ServerDispatchers into
AnnotatedServerDispatchers.

Solution:

Introduce the ServerDispatcher type, which has two methods dispatch
and handle. Introduce the AnnotatedServerDispatcher type as well,
which can take any ServerDispatcher, and return that dispatcher
with WireRecv on receipt and WireSend upon a succesful write
by the transporter.