finagle-core: implementing RequestMeterFilter based on new AsyncMeter
Problem

We wanted to be able to control access to rate limited resources, by slowing down requests to fit on the request rate accepted by these resources.

Solution

We ended up implementing a filter named RequestMeterFilter using Twitter util's AsyncMeter to achieve that. Feedback on naming are really welcomed.

Result

As a result, I've added this Filter so other users with the same needs can use this solution.

Signed-off-by: Alex Leong <aleong@twitter.com>

RB_ID=777050