finagle-core: Increase the Penalty in p2cEwma to a higher value
Problem

The PeakEWMA load balancer applies a penalty to the load of a server when it
doesn't have any historical latency. That penalty is initialized at
Double.MaxValue/2, and we add the number of outstanding messages to compute
the load of a server. The idea behind that, is that the load balancer behaves
like LeastLoaded at start-up before having any historical data.

The problem is that the Penalty is a very big double, and adding a small
integer will not change its value because of the lack of double precision in
that range.

Solution

Choosing a bigger Penalty (e.g. Long.MaxValue >> 16 ~= 140737 seconds or 39 hours)
fixes the problem and also works well with potentially extremely latent systems.

Signed-off-by: Ruben Oanta <roanta@twitter.com>

RB_ID=780176
NO-QUEUE=true