[finagle-serversets] Remove newZk flag
Problem
--

zk2 has been enabled by default for more than a year. The old zk! scheme should not be used.

Solution
--

Remove the option to use the zk! scheme for wily paths and the serverset! scheme. Unfortunately
we need to keep the scheme around for the many places where zk! is used directly.

RB_ID=780169