finagle-mux: Fix synchronization in TagSet
Motivation

In the `iterator` on the `TagSet.apply` TagSet, the synchronization on
`TagSet.this` was using the outer `TagSet` companion object instead of
the anonymous class. This can be observed by examining the bytecode
output.

Solution

Use an explicit `self` reference for synchronization.

Result

Correct synchronization.

Snippet of `next()` bytecode before:
```
  6: getstatic     #450                // Field com/twitter/finagle/mux/util/TagSet$.MODULE$:Lcom/twitter/finagle/mux/util/TagSet$;
  9: dup
 10: astore_2
 11: monitorenter
```

Snippet of `next()` bytecode after:
```
  7: getfield      #446                // Field $outer:Lcom/twitter/finagle/mux/util/TagSet$$anon$1;
 10: dup
 11: astore_2
 12: monitorenter
```

RB_ID=785034