finagle: Enable with-prefixed API for all protocols
Problem

There is a new, user-friendly API with `with` methods which only available
for HTTP, but not the rest of Finagle protocols.

Solution

Enable the `with` API for the all the Finagle protocols.

The challenge here is to be able to exclude/include some group of params
depending on the protocol. It turns out there are two group of params in
Finagle, which are not included into its every protocol: session pooling
(default pool vs. singleton pool) and load balancing (dafault balancer vs.
concurrent balancer).

The following table illustates what params should be enabled/disabled for
a particular protocol:

+--------------------------------------+
| Protocol  | Pool        | Balancer   |
+--------------------------------------+
| Http      | default     | default    |
| Mux       | singleton   | default    |
| MySql     | default     | default    |
| Thrift    | default     | default    |
| ThriftMux | singleton   | default    |
| Redis     | singleton   | default    |
| Memcached | singleton   | concurrent |
+--------------------------------------+

To solve this problem in a generic way this RB introduces `WithX` traits
that contains a single `withX` method pointing to a proper configuration
(e.g. `ConcurrentLoadBalancingParams` vs. `DefaultLoadBalancingParams`).

In addition to ability to mix-in a required configuration to every client,
the `WithX` traits are also quite useful to inheric Scaladoc and not copy
same test all over the places.

RB_ID=789092