finagle-core - Fix issue in ServerBuilder around protocol library
Problem

When using the ServerBuilder there is a bug where the protocol
library is not correctly picked up from the Stack Params. The
means that the server is incorrectly registered in the server
registry availble via the admin interface.

Resolution

Fix the anonymous function that converts from Stack params to
a Server to correctly use the fully configured params when
determining if the protocol library param should be added to
the list of server params.

Result

The protocol library is correctly configured for services
using the ServerBuilder.

RB_ID=791538