finagle-mux: Interrupt server writes on Tdiscarded
Problem

When writing a response in mux, we lose track of the writer
promise (i.e. write and forget). This is unfortunate because
we want to attempt to interrupt the write if we receive a
cancellation from the client. For example, this becomes
increasingly important with large responses that are more likely
to timeout.

Solution

Compose writes as part of our pending state on the mux server.

RB_ID=792186