finagle-mux: ping failure detector to close session by default
Problem

After a session is marked busy in mux, it is stuck in the
busy state until the session is closed. This is problematic
when there is an alternative good networking path, but client
never has the chance to reestablish the session on the good
path.

Solution

Close the session by default when ping RTT excceeds 4 seconds,
to allow a new session to be esbalished.

RB_ID=773649