[finagle-core] Make FixedInetResolver and 1 API public
[problem]

Though the InetResolver is public, the FixedInetResolver (which caches all successful lookups) is private[finagle]

[solution]

Make the FixedInetResolver have the same access level as the InetResolver. Also, make toAddr (the API which does the one-time lookup outside of polling) a separate public API for others to leverage as needed. Nothing new is being made public API-wise

RB_ID=794785