finagle-core: Expose a server's actual bound port in its ServerRegistry entry
Problem

We currently register a server being served in the ServerRegistry with the
addr requested and not the actual bound addr. This makes the registry somewhat
less than useful when using ephemeral ports as we end up reporting ":0" as
opposed to the bound address. The registry is then able to accurately report
a running server's bound port.

Resolution

Update ServerBuilder#serve to register the server with the
underlying.boundAddress which is the actual bound port of the server.

Result

Better information reported in the ServerRegistry that can be used to query
for a running server's port.

RB_ID=795441