finagle-core: Skip TimeoutFilter when possible
Motivation

Clients add a `c.t.f.service.TimeoutFilter` to their stack of filters
even when the supplied duration does not make sense. This is a waste
of resources and can be easily avoided, much as was being done for
servers.

Solution

Avoid adding the filter when the duration is not finite or if is not
positive.

While working on this, I also fixed up the server's name that is
passed into `SourceTrackingMonitor` as well as touched up the output
from said monitor.

RB_ID=795936