finagle-http: Add extractors for Status of HTTP Response
Problem

When writing response classifiers, and other code that deals with responses,
in a generic manner, it is currently inconvenient to match against specific
ranges of errors such as Server Errors or Client errors.

Solution

By adding extractors for the different categories (Unknown, Informational,
Successful, Redirection, Client Error, and Server Error) it becomes possible
to easily determine in which category a response falls.

Result

Determining whether a response was succesful is as easy as matching
on Successful(_).