finagle-core: StatsFilter: Use a synthetic exception for non-Exception failures
Motivation

The `ResponseClassification` code now allows `Returns` to be
classified as failures. The current code in `StatsFilter` does not
count these as failures in its `ExceptionStatsHandler` leading to some
confusion.

Solution

Create a synthetic exception, `ResponseClassificationSyntheticException`,
that is used when non-`Throws` are classified as failures.

Result

More accurate metrics.

RB_ID=797433