Motivation:
Clients annotate WireSend and WireRecv at the bottom of the finagle
stack. There is no parallel for a Server.

Solution:

Introduce the ServerDispatcher type, which has two methods dispatch
and handle. Introduce the ServerDispatcherInitializer as well,
which inits Tracing in a Server Dispatcher outside of the stack.

Result:

All dispatcher utilizing the SerialServerDispatcher now record
WireRecv and WireSend. Other dispatchers may utilize the
ServerDispatcherInitializer to record as appropriate.