finagle-core: Handle conflicts between service failures and policy in failure accrual
Problem
--

When FailureAccrualFactory is in probing state and a failing response is
recorded, we rely on the FailureAccrualPolicy to decide when to retry. However,
if the policy decides that we are healthy (for instance if we are above success
rate for the current window), we stay stucked in ProbeClosed state and never
retry. In other words, FailureAccrualFactory assumes that the policy internal
state cannot improve on failure, but has no way to ensure that this invariant
holds.

Solution
--

When the underling service fails but the policy indicates that we should not
mark dead, move back to probing state.

Result
--

FailureAccrualFactory cannot get stucked in ProbeClosed state, regardless of the
policy.

RB_ID=797807