finagle-mux: handshakes
Problem

We need to be able to iterate on our mux implementations in a
principled way. Session negotiation and versioning is supported
by the mux spec, but is not part of our implemention.

Solution

Introduce session handshaking for both clients and servers.

The mux spec allows for (re)negotiation to happen arbitrarily through
out a session, but for simplicity, our implementation assumes it
at the start of a session. It is implemented in terms of a Transport
so that negotiation can sit transparently below client and server
dispatchers and easily install features based on the exchanged
version and headers.

Unfortunately, we didn't ship Mux with handshakes, so we need to jump
through some hoops for backwards compatibility. We implement a
"pre-handshake" which checks whether or not a server can understand
the Tinit message. This allows us to roll out the patch without coordinating
updates between servers and clients.

RB_ID=793350