finagle: Include upstream/downstream addresses/client ids and trace id in exceptions
Problem

When a server throws an error on a request, there isn't an easy way to
determine where the request came from.

Solution

Include upstream/downstream addresses/client ids and trace id in exceptions.
The upstream address is propagated through a local context from the server
to the client.

RB_ID=797082
TBR=true