finagle-core: TrafficDistributor status becomes best of all weight classes
Problem

The retry module only retries when the underlying
ServiceFactory is in an Open status. Currently
TrafficDistributor becomes busy if any weight classes
is busy. This prevent retryable failures from being retried.

Solution

Change the status of TrafficDistributor to be the
best of all weight classes. Added status stats in the
retry module to make the problem easier to detect.

RB_ID=803588