Make MapHeaderMap completely case insensitive
Problem

MapHeaderMap class which is used for holding http headers is supposed
to work in a case insensitive manner. However some methods defined in
this class are not obeying this contract.

Solution

The proposed changes includes refactoring of the methods mentioned
above. The interface of the class has not changed. The logic of how
headers are stored in the underlying object has been modified to
provide functionality in an case insensitive manner.

Result

All the methods of MapHeaderMap works in a case insensitive manner.