Problem
While laying the groundwork for moving finagle-mux to netty4, it
became apparent that c.t.io.Buf has a clumsy API for decoding
numeric values. For example, the Mux protocol makes heavy use of
length encoded fields. So to decode a field we need to do the
following:

```
val values = ArrayBuffer()
var b0 = buf
while (b0.length > 0) {
  val Buf.U32BE(len, b1) = b0
  val field = b1.slice(0, len)
  b0 = b1.slice(len, b1.length)
  values += field
}
```

Compare this with a stateful API like Netty's ChannelBuffer:

```
val values = ArrayBuffer()
while (cb.readableBytes() > 0) {
  val field = cb.readSlice(cb.readInt())
  values += field
}
```

What's more concerning though is the performance profile. Because
Bufs can only be created dynamically via concat, this results in a
suboptimal memory layout (i.e. we forgo the optimizations that come
with sequential array access). This is apparent in a simple benchmark:
https://gist.github.com/roanta/6f024386c270dc5e9d66

Solution

Introduce a stateful API for constructing a Buf and extracting
numeric values.  The API is simple and currently only has an
implementation for writing and reading big-endian numeric values.
It's private finagle and we can decide to promote it to util as we
start porting more of the protocols and settle on a stable API.

Result

Decoding (and encoding) becomes much more natural to express:

```
val values = new ArrayBuffer[Buf]
val br = BufReader(buf)
while (br.remaining > 0) {
  val v = br.readBytes(br.readIntBE())
  values += v
}
```

and the performance is quite nice across the board:

[info] BufCodecBenchmark.decodeBuf                                         100  avgt   10    57.431 ±   4.626   ns/op
[info] BufCodecBenchmark.decodeBuf:·gc.alloc.rate.norm                     100  avgt   10   272.003 ±   0.001    B/op

[info] BufCodecBenchmark.decodeCB                                          100  avgt   10   964.125 ±  71.177   ns/op
[info] BufCodecBenchmark.decodeCB:·gc.alloc.rate.norm                      100  avgt   10  5048.054 ±   0.006    B/op

[info] BufCodecBenchmark.encodeBuf                                         100  avgt   10  1698.580 ±  55.522   ns/op
[info] BufCodecBenchmark.encodeBuf:·gc.alloc.rate.norm                     100  avgt   10  3376.092 ±   0.009    B/op

[info] BufCodecBenchmark.encodeCB                                          100  avgt   10  1211.978 ±  16.924   ns/op
[info] BufCodecBenchmark.encodeCB:·gc.alloc.rate.norm                      100  avgt   10   952.063 ±   0.007    B/op

RB_ID=801488