Problem
In some cases, when receiving a message which can't be translated to thrift,
the downgrader would echo an Rerr message back into the dispatcher. This isn't
exactly to spec and there are better ways to handle this case. What's worse,
this could cause an infinite recursion where we keep sending `Rerr`s between
the dispatcher and the syntethic client.

Solution

Uniformly handle errors by throwing in the pipeline. This will be handled by
Finagle's server bridge and be logged.

RB_ID=805589