finagle-core: FailureAccrualPolicy is mutable and should never be shared
Problem

`FailureAccrualPolicy` maintains a mutable state that `FailureAccrualFactory` uses
to determine whether or not an endpoint should be marked unavailable. Given that
it's mutable, it should never be shared across different endpoint stacks (endpoints)
in a single Finagle client. This is why using `Param(p: FailureAccrualPolicy)` may
lead to a surprising results when different endpoints interfere with each other.

Solution

`FaulureAccrualPolicy` should be passed as a function (factory) instead.

RB_ID=802953
TBR=true