finagle-mux: proxy peerCert during handshake
Problem

We capture the peerCert inside a local when constructing
a dispatcher. However, we defer the resolution of it until
we are done handshaking in `DeferredTransport`. We have access
to it before the handshake and it isn't likely to change.

Solution

Proxy the synchronous methods while we are handshaking.

RB_ID=809235