Implement server-side request entity decoding (#476 & #469).
Ideally, I'd have also implemented client-side encoding, but this is currently problematic (see https://github.com/netty/netty/issues/4970).
This also means that current tests, while not bad, are not as exhaustive as they could have been. Should client side encoding ever be implemented, it is suggested to drop `ServerSideDecodingTest` entirely and replace it with a dedicated entry in `EndToEndTest`.